fix: Tiers types
