+++ src/modules/common/types.ts
@@ -34,7 +34,7 @@ import { UIState } from 'modules/ui/reducer'
 import { UndoableSceneState } from 'modules/scene/reducer'
 import { EntityState } from 'modules/entity/reducer'
 import { CurationState } from 'modules/curation/reducer'
-import { TiersState } from 'modules/tiers/types'
+import { TiersState } from 'modules/tiers/reducer'
 
 const storageLoad = () => action(STORAGE_LOAD, {} as RootState)
 export type StorageLoadAction = ReturnType<typeof storageLoad>
